rad: remove old source-term files, complete reimplementation
Delete source_terms_multi_group.hpp, source_terms_single_group.hpp, and
radiation_dust_system.hpp. Move ComputeModelDependentKappaEAndKappaP and
ComputeModelDependentKappaFAndDeltaTerms implementations into
opacity_evaluation.hpp, and AddSourceTermsSingleGroup into source_terms.hpp.
Remove stale declarations (UpdateFlux, SolveLinearEqs, SolveLinearEqsWithLastColumn,
ComputeJacobianForGas*, SolveGas*RadiationEnergyExchange, AddSourceTermsMultiGroup)
from radiation_system.hpp. Fix RadMarshakDust include to use radiation_system.hpp.
