+++ src/problems/RadMarshakDust/testRadMarshakDust.cpp
@@ -13,7 +13,7 @@
 #include "AMReX.H"
 #include "QuokkaSimulation.hpp"
 #include "physics_info.hpp"
-#include "radiation/radiation_dust_system.hpp"
+#include "radiation/radiation_system.hpp"
 #include "util/BC.hpp"
 #include "util/fextract.hpp"
 #include "util/valarray.hpp"
